Revamp code actions implementations
